Add CMSeFracWare1DInstaller_web.exe
This does not require the user to have Matlab installed. Running the file will launch an application installer for Windows systems and, if the user doesn’t have it, it will automatically install the Matlab runtime engine. After installation, the user can run the software in the same way as any other Windows application.